flow_meter: changed counter variable type in pstats plugin

+++ flow_meter/pstatsplugin.h
@@ -70,7 +70,7 @@ struct RecordExtPSTATS : RecordExt {
    uint8_t pkt_tcp_flgs[PSTATS_MAXELEMCOUNT];
    struct timeval pkt_timestamps[PSTATS_MAXELEMCOUNT];
    int8_t pkt_dirs[PSTATS_MAXELEMCOUNT];
-   uint8_t pkt_count;
+   uint16_t pkt_count;
 
    typedef enum eHdrFieldID
    {
@@ -140,7 +140,7 @@ struct RecordExtPSTATS : RecordExt {
       ur_array_allocate(tmplt, record, F_PPI_PKT_FLAGS, pkt_count);
       ur_array_allocate(tmplt, record, F_PPI_PKT_DIRECTIONS, pkt_count);
 
-      for (uint8_t i = 0; i < pkt_count; i++) {
+      for (int i = 0; i < pkt_count; i++) {
          ur_time_t ts = ur_time_from_sec_usec(pkt_timestamps[i].tv_sec, pkt_timestamps[i].tv_usec);
          ur_array_set(tmplt, record, F_PPI_PKT_TIMES, i, ts);
          ur_array_set(tmplt, record, F_PPI_PKT_LENGTHS, i, pkt_sizes[i]);
